move prompt_for_password_and_derive_kek to encryption module
